Refactoring ProductionLogStepDTO to LogStepCacheDTO and Moving the StepAttemptCacheDTO to its proper location
