+++ MESS/MESS.Data/DTO/ProductionLogDTOs/LogSteps/LogStepCacheDTO.cs
@@ -1,10 +1,12 @@
-﻿namespace MESS.Data.DTO.ProductionLogDTOs;
+﻿using MESS.Data.DTO.ProductionLogDTOs.LogSteps.Attempts;
+
+namespace MESS.Data.DTO.ProductionLogDTOs.LogSteps;
 
 /// <summary>
 /// Represents a step in the production log. Used primarily for client-side caching purposes.
 /// Now supports multiple attempts per step.
 /// </summary>
-public class ProductionLogStepDTO
+public class LogStepCacheDTO
 {
     /// <summary>
     /// Gets or sets the ID of the work instruction step.
@@ -19,5 +21,5 @@ public class ProductionLogStepDTO
     /// <summary>
     /// Gets or sets the collection of attempts for this step.
     /// </summary>
-    public List<StepAttemptCacheDTO> Attempts { get; set; } = new();
+    public List<StepAttemptCacheDTO> Attempts { get; init; } = [];
 }
